requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
fastapi==0.129.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.63.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
lxml==6.0.2
MarkupSafe==3.0.3
openai==2.21.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
pypdfium2==5.4.0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
requests==2.32.5
sarvamai==0.1.25
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
